Move `ValidationErrors` to excs.py and give it a superclass
