+++ tests/test_param_validation.py
@@ -1,7 +1,7 @@
 import pytest
 
 from valohai_yaml.objs import Parameter
-from valohai_yaml.validation import ValidationErrors
+from valohai_yaml import ValidationErrors
 
 
 @pytest.mark.parametrize('case', (-5, 15))
