+++ valohai_yaml/__init__.py
@@ -1,4 +1,5 @@
 from .parsing import parse  # noqa
-from .validation import validate, ValidationErrors  # noqa
+from .validation import validate  # noqa
+from .excs import ValidationErrors  # noqa
 
 __version__ = '0.14.1'
