+++ valohai_yaml/objs/parameter.py
@@ -2,7 +2,7 @@ from enum import Enum
 from typing import List, Optional, Union
 
 from valohai_yaml.utils import listify
-from valohai_yaml.validation import ValidationErrors
+from valohai_yaml.excs import ValidationErrors
 from .base import Item
